Re-add logic to space out duplicate items
I think duplicate important items is a good solution to the "100%ing every seed"-problem with the full shuffle, and I think it's a lot better if the items aren't in the same place. I can even see this being enabled for Max Rando, and I don't want it seeming jank.