Add in better NES RNG [work-in-progress]
The vanilla game's RNG, in many cases, outputs poorly randomized values that form too much of a pattern. I've inserted a better NES RNG to see if that makes things more fun.

This can pretty much replace the need for "Reduce Dripper Variance". In theory it should also reduce the need for the 50/50 Alternating Jar, but that feature seems well-liked so probably can't remove that.

This assumes $F9-$FB in zero page memory is unused - which seems to be true so far in my testing.

Things I specifically don't think should be changed to use this as players are used to how they work:
- Overworld enemy random movement
- Bot/Bit/Myu random movement

Other enemies RNG behavior is a maybe for now.

Potential future things to do:
- Rework Standardize Drops to use a separate seed in SRAM (that will start the same for all players) and is used for drops only, removing the predictability of the current Standardize Drops solution.